Split slicer.rs into slicer/ submodule directory
types.rs: SliceMeshBuffers, uniform structs, constants
pipeline.rs: AdvancedSlicer struct and impl
tests.rs: WGSL shader validation tests
